cleanup repo
