--- src/napari_feature_classifier/Annotations.csv
@@ -1,4 +1,0 @@
-Label,annotations
-1,4.0
-2,2.0
-4,1.0
