--- src/napari_feature_classifier/Classifier_output.csv
@@ -1,4 +1,0 @@
-Label,annotations
-1,3
-2,2
-4,3
