--- src/napari_feature_classifier/annotation_lbls.csv
@@ -1,17 +1,0 @@
-,annotations,annotation_names
-1,1.0,early M
-2,2.0,late M
-3,3.0,early S
-4,1.0,early M
-5,2.0,late M
-6,3.0,early S
-7,,NoClass
-8,,NoClass
-9,1.0,early M
-10,3.0,early S
-11,3.0,early S
-12,1.0,early M
-13,2.0,late M
-14,2.0,late M
-15,,NoClass
-16,,NoClass
