--- src/napari_feature_classifier/annotation_lbls2.csv
@@ -1,17 +1,0 @@
-,annotations,annotation_names
-1,4.0,mid S
-2,3.0,early S
-3,5.0,late S
-4,4.0,mid S
-5,3.0,early S
-6,5.0,late S
-7,,NoClass
-8,,NoClass
-9,,NoClass
-10,5.0,late S
-11,,NoClass
-12,4.0,mid S
-13,3.0,early S
-14,,NoClass
-15,,NoClass
-16,,NoClass
