+++ src/napari_feature_classifier/napari.yaml
@@ -8,23 +8,18 @@ contributions:
     - id: napari-feature-classifier.load_classifier
       python_name: napari_feature_classifier.classifier_widgets:load_classifier
       title: Load a Classifier
-    - id: napari-feature-classifier.annotator_widget_v2
-      python_name: napari_feature_classifier.annotator_widget_v2:start_annotator
+    - id: napari-feature-classifier.annotator_widget
+      python_name: napari_feature_classifier.annotator_widget:LabelAnnotator
       title: Annotator
-    - id: napari-feature-classifier.test_widget
-      python_name: napari_feature_classifier.test_widget:test_widget
-      title: Test Widget
-    - id: napari-feature-classifier.annotator_widget_v3
-      python_name: napari_feature_classifier.annotator_widget_v3:LabelAnnotator
-      title: Test Widget
+    - id: napari-feature-classifier.annotator_init_widget
+      python_name: napari_feature_classifier.annotator_init_widget:InitializeLabelAnnotator
+      title: Init Annotator
   widgets:
     - command: napari-feature-classifier.initialize_classifier
       display_name: Initialize a Classifier
     - command: napari-feature-classifier.load_classifier
       display_name: Load Classifier
-    - command: napari-feature-classifier.test_widget
-      display_name: Test Widget
-    - command: napari-feature-classifier.annotator_widget_v2
-      display_name: Annotator_v2
-    - command: napari-feature-classifier.annotator_widget_v3
-      display_name: Annotator_v3
+    - command: napari-feature-classifier.annotator_widget
+      display_name: Annotator
+    - command: napari-feature-classifier.annotator_init_widget
+      display_name: Init Annotator
