--- src/napari_feature_classifier/test.csv
@@ -1,60 +1,0 @@
-filename_prefix,Label,annotations_x,annotations_y
-20211119_ABpanelTestset_2pair_3h_1_s2,1,1,1
-20211119_ABpanelTestset_2pair_3h_1_s2,2,2,1
-20211119_ABpanelTestset_2pair_3h_1_s2,3,1,1
-20211119_ABpanelTestset_2pair_3h_1_s2,4,1,1
-20211119_ABpanelTestset_2pair_3h_1_s2,5,1,1
-20211119_ABpanelTestset_2pair_3h_1_s2,6,2,2
-20211119_ABpanelTestset_2pair_3h_1_s2,7,2,2
-20211119_ABpanelTestset_2pair_3h_1_s2,9,2,2
-20211119_ABpanelTestset_2pair_3h_1_s2,10,2,2
-20211119_ABpanelTestset_2pair_3h_1_s2,11,2,2
-20211119_ABpanelTestset_2pair_3h_1_s2,12,2,2
-20211119_ABpanelTestset_2pair_3h_1_s2,13,2,2
-20211119_ABpanelTestset_2pair_3h_1_s2,15,2,2
-20211119_ABpanelTestset_2pair_3h_1_s2,18,2,2
-20211119_ABpanelTestset_2pair_3h_1_s2,22,2,2
-20211119_ABpanelTestset_2pair_3h_1_s2,23,2,
-20211119_ABpanelTestset_2pair_3h_1_s2,25,1,
-20211119_ABpanelTestset_2pair_3h_1_s2,28,2,
-20211119_ABpanelTestset_2pair_3h_1_s2,29,1,
-20211119_ABpanelTestset_2pair_3h_1_s2,31,2,
-20211119_ABpanelTestset_2pair_3h_1_s2,32,1,
-20211119_ABpanelTestset_2pair_3h_1_s2,33,1,
-20211119_ABpanelTestset_2pair_3h_1_s2,34,1,
-20211119_ABpanelTestset_2pair_3h_1_s2,35,2,
-20211119_ABpanelTestset_2pair_3h_1_s2,36,2,
-20211119_ABpanelTestset_2pair_3h_1_s2,37,2,
-20211119_ABpanelTestset_2pair_3h_1_s2,38,2,
-20211119_ABpanelTestset_2pair_3h_1_s2,39,1,
-20211119_ABpanelTestset_2pair_3h_1_s2,40,1,
-20211119_ABpanelTestset_2pair_3h_1_s2,41,1,
-20211119_ABpanelTestset_2pair_3h_1_s2,42,1,
-20211119_ABpanelTestset_2pair_3h_1_s2,43,1,
-20211119_ABpanelTestset_2pair_3h_1_s2,44,1,
-20211119_ABpanelTestset_2pair_3h_1_s2,45,2,
-20211119_ABpanelTestset_2pair_3h_1_s2,46,1,
-20211119_ABpanelTestset_2pair_3h_1_s2,47,1,
-20211119_ABpanelTestset_2pair_3h_1_s2,48,1,
-20211119_ABpanelTestset_2pair_3h_1_s2,49,2,
-20211119_ABpanelTestset_2pair_3h_1_s2,50,2,
-20211119_ABpanelTestset_2pair_3h_1_s2,51,2,
-20211119_ABpanelTestset_2pair_3h_1_s2,53,1,
-20211119_ABpanelTestset_2pair_3h_1_s2,54,2,
-20211119_ABpanelTestset_2pair_3h_1_s2,55,1,
-20211119_ABpanelTestset_2pair_3h_1_s2,56,2,
-20211119_ABpanelTestset_2pair_3h_1_s2,57,1,
-20211119_ABpanelTestset_2pair_3h_1_s2,58,1,
-20211119_ABpanelTestset_2pair_3h_1_s2,59,2,
-20211119_ABpanelTestset_2pair_3h_1_s2,60,1,
-20211119_ABpanelTestset_2pair_3h_1_s2,61,1,
-20211119_ABpanelTestset_2pair_3h_1_s2,62,1,
-20211119_ABpanelTestset_2pair_3h_1_s2,63,2,
-20211119_ABpanelTestset_2pair_3h_1_s2,64,2,
-20211119_ABpanelTestset_2pair_3h_1_s2,65,1,
-20211119_ABpanelTestset_2pair_3h_1_s2,66,1,
-20211119_ABpanelTestset_2pair_3h_1_s2,67,2,
-20211119_ABpanelTestset_2pair_3h_1_s2,68,1,
-20211119_ABpanelTestset_2pair_3h_1_s2,69,1,
-20211119_ABpanelTestset_2pair_3h_1_s2,70,1,
-20211119_ABpanelTestset_2pair_3h_1_s2,71,2,
